dlc changes

+++ Skyve.Systems.CS2/Domain/Api/PackageData.cs
@@ -19,7 +19,7 @@ public class PackageData
 	public PackageStability Stability { get; set; }
 	public PackageUsage Usage { get; set; } = (PackageUsage)(-1);
 	public PackageType Type { get; set; }
-	public List<uint> RequiredDLCs { get; set; }
+	public List<ulong> RequiredDLCs { get; set; }
 	public List<string> Tags { get; set; }
 	public List<PackageLink> Links { get; set; }
 	public List<PackageStatus> Statuses { get; set; }
